Refactor guide into separate component
